move back into a single file; revert typing changes; use regex better
